nl2sql_agent/app/requirements.txt
google-adk==0.1.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.89.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.11.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.1
pandas
# sentence_transformers
# pickle5
# faiss-cpu
# hf_xet
openai
numpy  
pymongo
